consolidate features/manageVault/ManageVaultDaiAllowance.tsx into features/openMultiplyVault/common/ManageVaultDaiAllowance.tsx
